MAINT: Define horizons in individual stratigraphic framework
