+++ frontend/src/components/project/stratigraphicFramework/StratigraphicFramework.tsx
@@ -1,38 +1,14 @@
-import { Button, Tooltip } from "@equinor/eds-core-react";
 import type React from "react";
 import { useMemo } from "react";
-import { Fragment } from "react/jsx-runtime";
 
 import type { RmsHorizon, RmsStratigraphicZone } from "#client";
 import { FrameworkDataContext } from "./FrameworkData";
 import { getZoneGridPlacement } from "./functions";
 import {
-  GridLine,
-  HorizonItem,
   StratigraphicFrameworkContainer,
   StratigraphicFrameworkContent,
   StratigraphicFrameworkHeader,
 } from "./StratigraphicFramework.style";
-
-function HorizonTooltipContent(
-  horizon: RmsHorizon,
-  isOrphan: boolean,
-  isUsedByZone: boolean,
-) {
-  return (
-    <>
-      {horizon.name}
-      <br />
-      Type: {horizon.type}
-      <br />
-      {isOrphan
-        ? "Horizon does not exist in RMS"
-        : isUsedByZone
-          ? "Horizon is used by one or more zones"
-          : ""}
-    </>
-  );
-}
 
 export function StratigraphicFramework({
   horizons,
@@ -45,7 +21,7 @@ export function StratigraphicFramework({
   onHorizonClick,
   maxHeight,
   disablePointerEvents = false,
-  children,
+  children: [horizonsComponent, zonesComponent],
 }: {
   horizons: RmsHorizon[];
   zones: RmsStratigraphicZone[];
@@ -57,7 +33,7 @@ export function StratigraphicFramework({
   onHorizonClick?: (horizon: RmsHorizon, isUnselected: boolean) => void;
   maxHeight?: string;
   disablePointerEvents?: boolean;
-  children: React.ReactNode;
+  children: React.ReactNode[];
 }) {
   const frameworkData = useMemo(() => {
     const zoneGridPlacement = getZoneGridPlacement(zones, horizons);
@@ -111,50 +87,10 @@ export function StratigraphicFramework({
         $numStratColumns={frameworkData.numStratColumns}
         $numRows={frameworkData.numGridRows}
       >
-        {horizons.map((horizon, index) => {
-          const rowStart = index * 2 + 1;
-          const isOrphan = frameworkData.orphanHorizonNamesSet.has(
-            horizon.name,
-          );
-          const isUsedByZone = frameworkData.horizonsUsedByZones.has(
-            horizon.name,
-          );
-          const isUnselected = frameworkData.unselectedHorizonNamesSet.has(
-            horizon.name,
-          );
-
-          return (
-            <Fragment key={horizon.name}>
-              <HorizonItem style={{ gridRow: `${rowStart} / span 2` }}>
-                <Tooltip
-                  title={HorizonTooltipContent(horizon, isOrphan, isUsedByZone)}
-                >
-                  <Button
-                    className={
-                      isOrphan ? "orphan" : isUnselected ? "unselected" : ""
-                    }
-                    onClick={() => onHorizonClick?.(horizon, isUnselected)}
-                    variant="ghost"
-                    color={isOrphan ? "danger" : "primary"}
-                    disabled={isUsedByZone && !isOrphan}
-                  >
-                    {horizon.name}
-                  </Button>
-                </Tooltip>
-              </HorizonItem>
-
-              <GridLine
-                $lineStyle={
-                  horizon.type.startsWith("interpreted") ? "solid" : "dashed"
-                }
-                style={{ gridRow: rowStart }}
-              />
-            </Fragment>
-          );
-        })}
-
         <FrameworkDataContext value={frameworkData}>
-          {children}
+          {horizonsComponent}
+
+          {zonesComponent}
         </FrameworkDataContext>
       </StratigraphicFrameworkContent>
     </StratigraphicFrameworkContainer>
